symmetric market max debt per share handling
the current `main` will only re-add a pool to a market when it gets
pinged by a user in the market.

however, this can lead to a situation where a pool can be drained of its funds if
the market is able to control the `reportedDebt` through some
manipulation. to do this, the attacker would move the `reportedDebt`
such that it surpasses the limit set by the pool, and then "bump" the
pool, and then move back below the reportedDebt. After bumping again, The market will then
effectively take the difference between the reduced reported debt and
the market limit becuase.

To prevent this from happening, this PR changes it so that a pool can be both added and
removed from a market by the specified limit.

notes:
* found one bug in `hardhat-router` recent merge which caused fully
  qualified path to be submitted for a module for some reason, so I
added a check to use `_contractName` function to prevent it from
spitting out a FQN
* new internal structure in `Market` is now required because an
  additional property, `liquidityAmount` must be stored for when a
pool is re-added to a market. Otherwise, everything functions as before.