audit feedback: fix IssueUSDModule debt bug
in audit feedback submitted by iosiro, it is possible to

it turns out this issue was related to insufficient testing of the
system collateralization ratio check, which should have been deducting
the balance of the system from the available credit capacity.
thankfully, this was easily resolved from a solidity perspective, it was
just a missing calculation in one branch.

iosiro's provided test has been altered slightly and brought into the
codebase to ensure we have strong coverage of this particular bug in the
future.

cc @iosiro-security