audit feedback: better handling of temporary market overrun
prior to this PR, pools will fail to re-join a market automatically
after being bumped out. the behavior used to be waiting for a pool to come and "absorb"
the debt, but this is non-intuitive and causes problems if all the pools
are bumped out. Much better pattern is just not to allow any more debt
accumulation until after the accumulation starts going below the last
accumulated value, at which point we can start bumping back in whereever
the maxValuePerShare of the top pool is.

NOTE: at this point, we are basically giving users control over the
location of the `maxValuePerShare` value, which could potentially
introduce an attack vector through a precision attack (ex. estitng the
value to something really, really big). thoughts?