fix incorrect calculations in max value per share
recent revelations from our auditing peers have shown that the
`getSystemValuePerShare` function can be dramatically simplified and
improved to meet our needs.

As it turns out, the branching condition which only became active with
an unset systemMinValuePerShare was the most correct part of the
function, since it correctly recongized that the maxValue taken up by a
pool is `creditCapacity * maxValuePerShare / creditCapacity`, or just
`maxValuePerShare`. With this in mind, and of course by subtracting the
debt per share (which can be computed outside this function), we end up
with the correct available credit capacity!