add new security features
these are things we should definitely have lined up for mainnet launch.
in particular:

* interaction delay timeout for the account, which prevents user from
  withdrawing funds from the system in the case of an accounting error
which allows for them to withdraw more than they should. this should
theoretically allow for us to be able to recover from basically any
attack by preventing loss of funds
* add a feature to feature flags which allows for denying to all if an
  address is in a list. this will allow for feature similar to the
"SystemStatus" system of v2x, but with more granularity and improved
code maintainability
* create new feature flags which are set to allowed automatically:
  * createAccount
  * deposit
  * withdraw
  * mintUsd
  * burnUsd
  * liquidate
  * liquidateVault
  * depositMarketCollateral
  * withdrawMarketCollateral
  * depositMarketUsd
  * withdrawMarketUsd
  * claimRewards
  * delegateCollateral

the existance of these feature flags will allow for us to selectively
disable any of these features in the case of a discovered attack in
progress/potentially exploitable.