add check to verify that the initial proxy has not been deployed
it is important to make sure that a new version of the initial proxy
isnt being created at this point.

this is done by doing a simple string search after the deployment is
created. If it contains `exec: contract.InitialProxy`, that means that
the proxy is redeployed and we don't want that, so the build will fail
on whatever PR it is or whatever.