add additional documention to clarify `associateDebt`
If an account is not part of a distribution, the amount of debt assigned
to the account would increase as expected. However, if an account is
already part of a distribution, and has some existing debt shares in the
vault, their debt will first be decreased pro-rata before being
increased by the associated debt amount. This functionality may seem
confusing, but its actually correct outcome when considering the use
case.

this was raised as part of the OZ audit.