M-01 fix cannot set decay rate
add a function to `setDecayRate` on `SpotMarketFactory`

additionally, it was identified that if we do this, we need to keep a
leash on what the supported types of collateral token implementations
are. so, this PR also limits the `upgradeSynthImpl` function to only
upgrading to the *current* synth implementation

also, there wer emissing tests for `upgradeSynthImpl`, so those have
been added.