ability to pass runtime values into oracle manager nodes
they can be passed as an array of keys and values. a simple loop
looking over all the keys can find any necessary . an example of how this could be used is for an external node to check to see what the
volume of a particular order is, to modify the oracle's resulting value based on the calculation.