fix: Fixes margin modification on settlement and liquidation
Aside from some structural changes to minimise gas and also overcome
max local vars in stack, this change also involves fixes to margin
management.

Noteably, during debt calculations, we use different margin relative to
each position. Since getMarginUsd on the `accountId` _does not_ look at
fees incurred on either the liquidation (liqReward/keeperFee) or
settlement (keeper/order fees), an incorrect amount of debt was
deducted.

Additionally, in updateCollateralUsd also did not include fees in the
next open position when either deducting margin (in that it would deduct
less than it should) or attributing margin in the form of sUSD (in that
it was attributing more than it should), where the difference is in
orderFee + keeperFee.

Note this has _not_ been tested yet. Once tests are written, edge
cases tested, and numbers crunched we should be more confident this is a
working solution.

A few things still remain, there are still TODOs and lots of work in
reducing uint256/int256 to 128 in most areas. `validateTrade` now also
returns a struct instead of a tuple due to the calling function's stack
too deep but I wonder how impactful this is to gas. It's also
inconsistent with other validateXXX functions which bothers me.

More improvements to come.