fixes for reward manager to report correct shares
basically this whole change is to move rewards manager position updated hook to occur
after the position changes have occured. the data sent to the distributor
is still the data from when the distribution was prior to update, but now we send
the message after the change has occured so its possible for the rewards distributor
to query for the new value. This is required for SnapshotRewardsDistributor.

it slightly improves gas. but it modestly increases complexity as well.