SettlementHook interface changes and rm order before execution
This also removes a few legacy tests that we no longer need, changes the
onSettle cb to just be accountId, marketId, and price where price is the
Pyth price on settlement.

It adds a few additional comments, adds a test to assert the cb values
in onSettle, and removes an unused pnl var on settleOrder runtime.