legacymarket: allow migration of liquidatable accounts
we remove the code that allows for the legacymarket to "liquidate" accounts, and instead allow even liquidatable accounts to be migrated so that the v3 liquidation system can be
used instead.

this change works by removing a restriction on the v3 core system which prevented the assignment of debt which would cause an account to become liquidatable, a tactic that would
allow a market to mint effectively unlimited debt.